+++ service/detailed_record_service.go
@@ -49,7 +49,7 @@ func DetailedAdd(detail request.DetailedAdd) error {
 //实收明细详情
 func DetailsDetail(id int) (res response.DetailedList, err error) {
 	record := model.DetailedRecord{}
-	record.ID = uint(id)
+	record.ID = int64(id)
 	res, err = record.GetDetail()
 	if err != nil {
 		return
@@ -61,7 +61,7 @@ func DetailsDetail(id int) (res response.DetailedList, err error) {
 func IsBind(detail request.DetailedEdit)  error{
 	defer casbin.ClearEnforcer()
 	d := model.DetailedRecord{}
-	d.ID = uint(detail.ID)
+	d.ID = int64(detail.ID)
 	//逻辑
-	return d.Binding(uint(detail.ID),detail.OrderCode,detail.IsBind)
+	return d.Binding(detail.ID,detail.OrderCode,detail.IsBind)
 }
